v2.1.3 后台页面完成

+++ src/main/java/com/bookshop/controller/BookController.java
@@ -1,5 +1,6 @@
 package com.bookshop.controller;
 
+import com.bookshop.common.checkSession;
 import com.bookshop.common.responseFromServer;
 import com.bookshop.entity.Book;
 import com.bookshop.service.BookService;
@@ -10,6 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpSession;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -29,71 +31,94 @@ public class BookController {
 
     @RequestMapping("/getBookById")
     @ResponseBody
-    public responseFromServer getBookById(@RequestBody Integer bookId){
+    public responseFromServer getBookById(@RequestBody Integer bookId) {
         return bookService.getBook(bookId);
     }
 
     @RequestMapping("/getBook")
     @ResponseBody
-    public responseFromServer getBook(@RequestBody Book book){
+    public responseFromServer getBook(@RequestBody Book book) {
         return bookService.getBook(book.getBookId());
     }
 
     @RequestMapping("/searchBooksPage")
     @ResponseBody
-    public responseFromServer searchBooksPage(@RequestBody Map<String,Object> requestMap){
+    public responseFromServer searchBooksPage(@RequestBody Map<String, Object> requestMap) {
         return bookService.searchBooksPage(requestMap);
     }
 
     /*书本推荐暂时先返回所有的书籍*/
     @RequestMapping("/getRecommendedBooks")
     @ResponseBody
-    public responseFromServer getRecommendedBooks(@RequestBody Map<String,Object> requestMap){
+    public responseFromServer getRecommendedBooks(@RequestBody Map<String, Object> requestMap, HttpSession session) {
         return searchBooksPage(requestMap);
     }
 
     @RequestMapping("/getAllBooksPage")
     @ResponseBody
-    public responseFromServer getAllBooksPage(@RequestBody Map<String,Object> queryMap){
-        return bookService.getAllBooksPage(queryMap);
+    public responseFromServer getAllBooksPage(@RequestBody Map<String, Object> queryMap, HttpSession session) {
+        if (checkSession.checkManager(session))
+            return bookService.getAllBooksPage(queryMap);
+        else return responseFromServer.needLogin();
     }
 
 
     @RequestMapping("/insertBook")
     @ResponseBody
-    public responseFromServer insertBook(@RequestBody Book book){
-        return bookService.insertBook(book);
+    public responseFromServer insertBook(@RequestBody Book book, HttpSession session) {
+        if (checkSession.checkManager(session))
+            return bookService.insertBook(book);
+        else return responseFromServer.needLogin();
     }
 
     @RequestMapping("/updateBook")
     @ResponseBody
-    public responseFromServer updateBook(@RequestBody Book book){
-        return bookService.updateBook(book);
+    public responseFromServer updateBook(@RequestBody Book book, HttpSession session) {
+        if (checkSession.checkManager(session))
+            return bookService.updateBook(book);
+        else return responseFromServer.needLogin();
+
     }
 
     @RequestMapping("/updateBooks")
     @ResponseBody
-    public responseFromServer updateBooks(@RequestBody List<Book> books){
-        return bookService.updateBooks(books);
+    public responseFromServer updateBooks(@RequestBody List<Book> books, HttpSession session) {
+        if (checkSession.checkManager(session)) {
+            return bookService.updateBooks(books);
+        } else {
+            return responseFromServer.needLogin();
+        }
+
     }
 
 
     @RequestMapping("/deleteBookById")
     @ResponseBody
-    public responseFromServer deleteBookById(@RequestBody Integer bookId){
-        return bookService.deleteBookById(bookId);
+    public responseFromServer deleteBookById(@RequestBody Integer bookId, HttpSession session) {
+        if (checkSession.checkManager(session))
+            return bookService.deleteBookById(bookId);
+        else
+            return responseFromServer.needLogin();
     }
 
 
     @RequestMapping("/deleteBook")
     @ResponseBody
-    public responseFromServer deteleBook(@RequestBody Book book){
-        return bookService.deleteBook(book);
+    public responseFromServer deteleBook(@RequestBody Book book, HttpSession session) {
+        if (checkSession.checkManager(session)) {
+            return bookService.deleteBook(book);
+        } else {
+            return responseFromServer.needLogin();
+        }
     }
+
     @RequestMapping("/deleteBooks")
     @ResponseBody
-    public responseFromServer deleteBooks(@RequestBody List<Book> books){
-        return bookService.deleteBooks(books);
+    public responseFromServer deleteBooks(@RequestBody List<Book> books, HttpSession session) {
+        if (checkSession.checkManager(session))
+            return bookService.deleteBooks(books);
+        else
+            return responseFromServer.needLogin();
     }
 
 }
