+++ src/main/java/com/bookshop/controller/UserController.java
@@ -1,8 +1,10 @@
 package com.bookshop.controller;
 
+import com.bookshop.common.checkSession;
 import com.bookshop.common.responseFromServer;
 import com.bookshop.entity.User;
 import com.bookshop.service.UserService;
+import org.apache.tools.ant.taskdefs.condition.Http;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -25,14 +27,14 @@ import java.util.Map;
 @Controller
 @RequestMapping("/user")
 public class UserController {
-        @Resource
-        UserService userService;
+    @Resource
+    UserService userService;
 
     @RequestMapping("/checkLogin")
     @ResponseBody
-    public responseFromServer checkLogin(HttpSession session){
-        User user = (User)session.getAttribute("user");
-        if(user!=null&&user.getUserId()!=null){
+    public responseFromServer checkLogin(HttpSession session) {
+        User user = (User) session.getAttribute("user");
+        if (user != null && user.getUserId() != null) {
             user.setUserPassword("");
             return responseFromServer.success(user);
         }
@@ -42,44 +44,35 @@ public class UserController {
 
     @RequestMapping("/register")
     @ResponseBody
-    public responseFromServer register(@RequestBody User user, HttpSession session){
+    public responseFromServer register(@RequestBody User user, HttpSession session) {
         responseFromServer response = userService.insertUser(user);
-        if(response.isSuccess()){
-            session.setAttribute("user",(User)response.getData());
+        if (response.isSuccess()) {
+            session.setAttribute("user", (User) response.getData());
         }
         return response;
-/*        Map<String,Object> result = new HashMap<>();
-
-        try{
-            user = userService.insertUser(user);
-        }catch(UserNameIsUsedException e){
-            user = null;
-            result.put("error","用户名被占用");
-        }catch(Exception e){
-            user = null;
-            result.put("error","注册失败");
-        }
-        result.put("user",user);
-        return result;*/
     }
 
     @RequestMapping("/registerUsers")
     @ResponseBody
-    public responseFromServer registerUsers(@RequestBody List<User> users){
-        Map<String,Object> result = new HashMap<>();
-        List<User> errorUsers = new ArrayList<>();
-        for (User user:users) {
-            try{
-                userService.insertUser(user);
-            }catch(Exception e){
-                errorUsers.add(user);
+    public responseFromServer registerUsers(@RequestBody List<User> users, HttpSession session) {
+        if (checkSession.checkManager(session)) {
+            Map<String, Object> result = new HashMap<>();
+            List<User> errorUsers = new ArrayList<>();
+            for (User user : users) {
+                try {
+                    userService.insertUser(user);
+                } catch (Exception e) {
+                    errorUsers.add(user);
+                }
             }
+            if (errorUsers.size() > 0) {
+                result.put("error", "批量注册用户失败");
+                result.put("errorUsers", errorUsers);
+            }
+            return userService.insertUsers(users);
+        } else {
+            return responseFromServer.needLogin();
         }
-        if(errorUsers.size()>0){
-            result.put("error","批量注册用户失败");
-            result.put("errorUsers",errorUsers);
-        }
-        return userService.insertUsers(users);
     }
 
 
@@ -87,46 +80,50 @@ public class UserController {
     /*todo 删除账号密码校验*/
     @RequestMapping("/closeAccount")
     @ResponseBody
-    public responseFromServer closeAccount(@RequestBody  User user) {
-        return userService.deleteUser(user);
+    public responseFromServer closeAccount(@RequestBody User user, HttpSession session) {
+        if (checkSession.checkManager(session)) {
+            return userService.deleteUser(user);
+        } else {
+            return responseFromServer.needLogin();
+        }
     }
 
     @RequestMapping("/setsession")
     @ResponseBody
     public responseFromServer setsession(HttpSession session) {
-        session.setAttribute("mysession","fuckyou");
+        session.setAttribute("mysession", "fuckyou");
         return responseFromServer.success();
     }
 
     @RequestMapping("/testsession")
     @ResponseBody
     public responseFromServer testsession(HttpSession session) {
-        String user = (String)session.getAttribute("mysession");
+        String user = (String) session.getAttribute("mysession");
         return responseFromServer.success();
     }
 
 
     @RequestMapping("/sessiontest")
     @ResponseBody
-    public responseFromServer sessionTest(HttpSession session){
-        User nowUser = (User)session.getAttribute("user");
+    public responseFromServer sessionTest(HttpSession session) {
+        User nowUser = (User) session.getAttribute("user");
         return null;
     }
 
 
     @RequestMapping("/toLogin")
-    public String toLogin(){
+    public String toLogin() {
         return "login";
     }
 
     @RequestMapping("/login")
     @ResponseBody
-    public responseFromServer login(@RequestBody  User user,HttpSession session){
+    public responseFromServer login(@RequestBody User user, HttpSession session) {
         responseFromServer response = userService.login(user);
-        if(response.isSuccess()){
-            User nowUser = (User)response.getData();
+        if (response.isSuccess()) {
+            User nowUser = (User) response.getData();
             user.setUserPassword(null);
-            session.setAttribute("user",nowUser);
+            session.setAttribute("user", nowUser);
         }
         return response;
     }
@@ -134,10 +131,10 @@ public class UserController {
 
     @RequestMapping("/logOut")
     @ResponseBody
-    public responseFromServer logOut(HttpSession session){
-        if((User)session.getAttribute("user")==null){
+    public responseFromServer logOut(HttpSession session) {
+        if ((User) session.getAttribute("user") == null) {
             return responseFromServer.error("暂无登录信息");
-        }else{
+        } else {
             session.removeAttribute("user");
             return responseFromServer.success();
         }
@@ -145,63 +142,80 @@ public class UserController {
 
     @RequestMapping("/getUser")
     @ResponseBody
-    public responseFromServer searchUsersPage(@RequestBody User user){
-        return userService.getUser(user.getUserId());
+    public responseFromServer getUser(@RequestBody User user, HttpSession session) {
+        if (checkSession.checkManager(session))
+            return userService.getUser(user.getUserId());
+        else if (checkSession.check(session)) {
+            /*user只能获取自己的用户信息*/
+            if (((User) session.getAttribute("user")).getUserId().intValue() != user.getUserId().intValue()) {
+                return userService.getUser(user.getUserId());
+            } else
+                return responseFromServer.error("非法操作");
+        }
+        return responseFromServer.needLogin();
     }
-
 
 
     @RequestMapping("/searchUsersPage")
     @ResponseBody
-    public responseFromServer searchUsersPage(@RequestBody Map<String,Object> requestMap){
-        return userService.searchUsersPage(requestMap);
+    public responseFromServer searchUsersPage(@RequestBody Map<String, Object> requestMap, HttpSession session) {
+        if (checkSession.checkManager(session)) {
+            return userService.searchUsersPage(requestMap);
+        } else return responseFromServer.needLogin();
     }
 
     @RequestMapping("/getAllUsers")
     @ResponseBody
-    public responseFromServer getAllUsers(){
-        return userService.getAllUsers();
+    public responseFromServer getAllUsers(HttpSession session) {
+        if (checkSession.checkManager(session))
+            return userService.getAllUsers();
+        return responseFromServer.needLogin();
     }
 
     @RequestMapping("/getAllUsersPage")
     @ResponseBody
-    public responseFromServer getAllUsersPage(@RequestBody Integer startPage){
-        return userService.getAllUsersPage(startPage);
+    public responseFromServer getAllUsersPage(@RequestBody Integer startPage, HttpSession session) {
+        if (checkSession.checkManager(session))
+            return userService.getAllUsersPage(startPage);
+        return responseFromServer.needLogin();
     }
 
     @RequestMapping("/updateUser")
     @ResponseBody
-    public responseFromServer updateUser(@RequestBody User user){
-        return userService.updateUser(user);
-        /*if(!userService.updateUser(user)){
-            return "error";
-        }else{
-            return "success";
-        }*/
+    public responseFromServer updateUser(@RequestBody User user, HttpSession session) {
+        if (checkSession.checkManager(session))
+            return userService.updateUser(user);
+        return responseFromServer.needLogin();
     }
 
     @RequestMapping("/updateUsers")
     @ResponseBody
-    public responseFromServer updateUsers(@RequestBody List<User> users){
-        return userService.updateUsers(users);
-        /*返回影响行数*/
+    public responseFromServer updateUsers(@RequestBody List<User> users, HttpSession session) {
+        if (checkSession.checkManager(session))
+            return userService.updateUsers(users);
+        return responseFromServer.needLogin();
     }
 
     @RequestMapping("/deleteUser")
     @ResponseBody
-    public responseFromServer deleteUser(@RequestBody User user){
-        return userService.deleteUser(user);
-        /*返回删除的行数*/
-        /*if(userService.deleteUser(user)){
-            return "success";
-        }else{
-            return "error";
-        }*/
+    public responseFromServer deleteUser(@RequestBody User user, HttpSession session) {
+        if (checkSession.checkManager(session))
+            return userService.deleteUser(user);
+        else if (checkSession.check(session)) {
+            if (((User) session.getAttribute("user")).getUserId().intValue() == user.getUserId().intValue()) {
+                return userService.deleteUser(user);
+            } else {
+                return responseFromServer.error("非法操作");
+            }
+        }
+        return responseFromServer.error();
     }
 
     @RequestMapping("/deleteUsers")
     @ResponseBody
-    public responseFromServer deleteUsers(List<User> users){
-        return userService.deleteUsers(users);
+    public responseFromServer deleteUsers(List<User> users, HttpSession session) {
+        if (checkSession.checkManager(session))
+            return userService.deleteUsers(users);
+        else return responseFromServer.needLogin();
     }
 }
